more render (#13)
* add new dev version

* add error_handler

* add error_handler

* catch if import name is different than pip name

* lets hardcode pydantic 2.5.3 and see if that fixes

* Revert "lets hardcode pydantic 2.5.3 and see if that fixes"

This reverts commit def20b592f59e057844ce90dd8c94a28bbd0d8df.

* 0.2.3.dev4, actually its problem with kube_hunter, nvm

* maybe we need the new requirements

* woot client functions are done?

* fix tests

* 0.2.3.dev7, dont assume client function, must specify with --client

* 0.2.3.dev8 fix small bug client functions

* 0.2.3.dev9, add fallbacks if unrecognized version of OpenAPI present in API functions

* add pyjwt to requirements

* add ability to do multiple error handlers!

* dont require pyjwt anymore

* 0.2.4.dev2, replace hacky manual mimetypes with real mimetypes

* add polycustom

* multi webhook (#9)

* towards multiple webhook handlers simultaneously

* next

* update pyproject.toml

* 0.2.4.dev5, accept non-200 and non-201 status codes from polyCustom

* 0.2.4.dev6 bring python client in line with execute server and consider 3xx to be error

* 0.2.4.dev7, make TypedDict input and response types more picky if wrong TypedDict is used

* actually use args in webhook

* 0.2.4.dev8, fix webhook stopping text

* getting there

* 0.2.4.dev9, handle already connected error on windows

* 0.2.4.deva1, better user facing messages for webhooks and error handlers

* woops lets go version 10

* 0.2.4.dev11

* move polyCustom to own file

* onward

* fix it

* wip

* next

* there we go
